requirements.txt
six==1.10.0
Flask==0.11.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
configargparse==0.10.0
click==6.6
itsdangerous==0.24
peewee==2.8.1
wsgiref==0.1.2
geopy==1.11.0
s2sphere==0.2.4
PyMySQL==0.7.5
flask-cors==2.1.2
flask-compress==1.3.0
LatLon==1.0.1
git+https://github.com/PokemonGoMap/pgoapi.git#egg=pgoapi
xxhash
sphinx==1.4.5
sphinx-autobuild==0.6.0
recommonmark==0.4.0
sphinx_rtd_theme==0.1.9
requests==2.10.0
PySocks==1.5.6
git+https://github.com/ChrisTM/Flask-CacheBust.git@aa09ad861f104f987928fe9f5107ccfe0e4473c3#egg=flask_cachebust
protobuf_to_dict==0.1.0
paho-mqtt==1.2